chore: cleanup build script
